improve ui

+++ src/main/resources/static/script/app/controller.hotel.js
@@ -5,6 +5,9 @@ angular.module("controller.hotel", ["services", "ui.router", "ngFileUpload"])
 		$scope.data.goto = function goto(state) {
 			$state.go(constants.hotelsStatePrefix + "." + state);
 		}
+		$scope.data.toggleNavMenu = function() {
+			$scope.data.shrink = !$scope.data.shrink;
+		}
 	}])
     .controller('hotelListCtrl', ['$scope', '$state', 'apiService', 'alertManager', 'modals', function ($scope, $state, apiService, alertManager, modals) {
         function initHotelList() {
@@ -37,12 +40,16 @@ angular.module("controller.hotel", ["services", "ui.router", "ngFileUpload"])
         $scope.data.addHotel = function() {
         	$state.go('FeastBooking.hotels.newHotel');
         };
-        $scope.data.deleteHotels = function(ids) {
+        $scope.data.deleteHotels = function() {
         	var ids = $scope.data.hotelList.filter(function(hotel) {
         		return hotel.$$selected;
         	}).map(function(selected) {
         		return selected.id;
         	});
+        	if(!ids || ids.length === 0) {
+        		alertManager.addAlert("warning", "請選擇要刪除的酒店");
+        		return;
+        	}
         	apiService.deleteHotels({ ids: ids}).then(function() {
         		alertManager.addAlert("success", "成功刪除酒店");
         		$scope.data.hotelList = $scope.data.hotelList.filter(function(hotel) {
